Moved facets service and mapper to the API project
