+++ src/SFA.DAS.Commitments.Api.UnitTests/Mapping/Service/ApprenticeshipFilterService/WhenFilterApprenticeships.cs
@@ -3,19 +3,19 @@ using System.Collections.Generic;
 using System.Linq;
 
 using FluentAssertions;
+
 using NUnit.Framework;
 
+using SFA.DAS.Commitments.Api.Orchestrators.Mappers;
 using SFA.DAS.Commitments.Api.Types.Apprenticeship;
 using SFA.DAS.Commitments.Api.Types.Apprenticeship.Types;
 
-namespace SFA.DAS.Commitments.Application.UnitTests.Service.ApprenticeshipFilterService
+namespace SFA.DAS.Commitments.Api.UnitTests.Mapping.Service.ApprenticeshipFilterService
 {
-    using SFA.DAS.Commitments.Application.Services;
-
     [TestFixture]
     public class WhenFilterApprenticeships
     {
-        private ApprenticeshipFilterService _sut;
+        private Api.Orchestrators.Mappers.ApprenticeshipFilterService _sut;
 
         private List<Apprenticeship> _apprenticeships;
 
@@ -40,7 +40,7 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Service.ApprenticeshipFilter
                                                DataLockCourseTriaged = true
                                            }
                                    };
-            _sut = new ApprenticeshipFilterService(new FacetMapper());
+            _sut = new Api.Orchestrators.Mappers.ApprenticeshipFilterService(new FacetMapper());
         }
 
         [TestCase(Originator.Provider)]
