+++ src/SFA.DAS.Commitments.Api.UnitTests/Mapping/Service/Facets/WhenExtractingProviders.cs
@@ -2,14 +2,13 @@
 using System.Linq;
 
 using FluentAssertions;
-
 using NUnit.Framework;
 
+using SFA.DAS.Commitments.Api.Orchestrators.Mappers;
 using SFA.DAS.Commitments.Api.Types.Apprenticeship;
 using SFA.DAS.Commitments.Api.Types.Apprenticeship.Types;
-using SFA.DAS.Commitments.Application.Services;
 
-namespace SFA.DAS.Commitments.Application.UnitTests.Service.Facets
+namespace SFA.DAS.Commitments.Api.UnitTests.Mapping.Service.Facets
 {
     [TestFixture]
     public class WhenExtractingProviders
