+++ src/SFA.DAS.Commitments.Api.UnitTests/Mapping/Service/Facets/WhenExtractingRecordStatuses.cs
@@ -6,20 +6,17 @@ using Castle.Components.DictionaryAdapter;
 using FluentAssertions;
 using NUnit.Framework;
 
+using SFA.DAS.Commitments.Api.Orchestrators.Mappers;
 using SFA.DAS.Commitments.Api.Types.Apprenticeship;
 using SFA.DAS.Commitments.Api.Types.Apprenticeship.Types;
-using SFA.DAS.Commitments.Api.Types.DataLock.Types;
-using SFA.DAS.Commitments.Application.Services;
 
-namespace SFA.DAS.Commitments.Application.UnitTests.Service.Facets
+namespace SFA.DAS.Commitments.Api.UnitTests.Mapping.Service.Facets
 {
     [TestFixture]
     public class WhenExtractingRecordStatuses
     {
         private List<Apprenticeship> _data;
-
         private ApprenticeshipSearchQuery _userQuery;
-
         private FacetMapper _sut;
 
         [SetUp]
