+++ src/SFA.DAS.Commitments.Api/SFA.DAS.Commitments.Api.csproj
@@ -302,8 +302,10 @@
     <Compile Include="Models\SystemDetails.cs" />
     <Compile Include="Orchestrators\ApprenticeshipOrchestrator.cs" />
     <Compile Include="Orchestrators\EmployerOrchestrator.cs" />
+    <Compile Include="Orchestrators\Mappers\ApprenticeshipFilterService.cs" />
     <Compile Include="Orchestrators\Mappers\ApprenticeshipMapper.cs" />
     <Compile Include="Orchestrators\Mappers\CommitmentMapper.cs" />
+    <Compile Include="Orchestrators\Mappers\FacetMapper.cs" />
     <Compile Include="Orchestrators\Mappers\IApprenticeshipMapper.cs" />
     <Compile Include="Orchestrators\Mappers\ICommitmentMapper.cs" />
     <Compile Include="Orchestrators\Mappers\IDataLockMapper.cs" />
