+++ src/SFA.DAS.Commitments.Application.UnitTests/SFA.DAS.Commitments.Application.UnitTests.csproj
@@ -169,12 +169,6 @@
     <Compile Include="Rules\ApprenticeshipUpdateRules\WhenDetermineNewAgreementStatus.cs" />
     <Compile Include="Rules\ApprenticeshipUpdateRules\WhenDetermineWhetherChangeRequireAgreement.cs" />
     <Compile Include="Rules\CommitmentRules\WhenDetermineAgreementStatus.cs" />
-    <Compile Include="Service\ApprenticeshipFilterService\WhenPagingFilterResults.cs" />
-    <Compile Include="Service\Facets\WhenExtractingApprenticeshipStatuses.cs" />
-    <Compile Include="Service\Facets\WhenExtractingProviders.cs" />
-    <Compile Include="Service\Facets\WhenExtractingRecordStatuses.cs" />
-    <Compile Include="Service\Facets\WhenExtractingTrainingCourse.cs" />
-    <Compile Include="Service\ApprenticeshipFilterService\WhenFilterApprenticeships.cs" />
     <Compile Include="StubCurrentDateTime.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -184,6 +178,10 @@
     </None>
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\SFA.DAS.Commitments.Api\SFA.DAS.Commitments.Api.csproj">
+      <Project>{3CB1D817-D35A-4EFE-8B93-5FF0FAB94F82}</Project>
+      <Name>SFA.DAS.Commitments.Api</Name>
+    </ProjectReference>
     <ProjectReference Include="..\SFA.DAS.Commitments.Application\SFA.DAS.Commitments.Application.csproj">
       <Project>{880CDAF6-D9EE-481B-B4D8-7DE9C215A770}</Project>
       <Name>SFA.DAS.Commitments.Application</Name>
