+++ src/SFA.DAS.Commitments.Application/SFA.DAS.Commitments.Application.csproj
@@ -191,8 +191,6 @@
     <Compile Include="Queries\QueryResponse{T}.cs" />
     <Compile Include="Rules\ApprenticeshipUpdateRules.cs" />
     <Compile Include="Rules\ICommitmentRules.cs" />
-    <Compile Include="Services\ApprenticeshipFilterService.cs" />
-    <Compile Include="Services\FacetMapper.cs" />
     <Compile Include="Services\HistoryService.cs" />
   </ItemGroup>
   <ItemGroup>
